+++ tests/nv/test_ex_routes.py
@@ -15,6 +15,8 @@
 # You should have received a copy of the GNU General Public License
 # along with NeoVintageous.  If not, see <https://www.gnu.org/licenses/>.
 
+import re
+
 import unittest
 
 from NeoVintageous.nv.ex.scanner import _ScannerState
@@ -28,6 +30,7 @@ from NeoVintageous.nv.ex_routes import _ex_route_noremap
 from NeoVintageous.nv.ex_routes import _ex_route_only
 from NeoVintageous.nv.ex_routes import _ex_route_substitute
 from NeoVintageous.nv.ex_routes import _ex_route_tabnext
+from NeoVintageous.nv.ex_routes import ex_routes
 from NeoVintageous.nv.ex_routes import TokenCommand
 from NeoVintageous.nv.ex_routes import TokenEof
 
@@ -282,7 +285,7 @@ class Test_ex_route_substitute(unittest.TestCase):
         )
 
 
-class Test_ex_route_tab_next(unittest.TestCase):
+class Test_ex_route_tabnext(unittest.TestCase):
 
     def test_can_scan(self):
         actual = _ex_route_tabnext(_ScannerState(''))
@@ -293,3 +296,154 @@ class Test_ex_route_tab_next(unittest.TestCase):
 
         actual = _ex_route_tabnext(_ScannerState(''))
         self.assertEqual(actual, (None, [TokenCommand('tabnext', forced=False), TokenEof()]))
+
+
+class TestRoutes(unittest.TestCase):
+
+    def routeMatch(self, string):
+        matches = []
+        for route, command in ex_routes.items():
+            match = re.compile(route).match(string)
+            if match:
+                matches.append((match, route, command))
+
+        return matches
+
+    def test_invalid_routes(self):
+        def assert_no_route(string):
+            self.assertEquals(self.routeMatch(string), [], 'failed asserting no route for {}'.format(string))
+
+        assert_no_route('')
+        assert_no_route('zfoobar')
+        assert_no_route(' ')
+        assert_no_route('$')
+
+        # TODO assert_no_route('foobar') shouldn't match :file
+
+    def test_routes(self):
+
+        def assert_route(string, first=False, name=None):
+            matches = self.routeMatch(string)
+
+            if not first:
+                self.assertEqual(
+                    len(matches), 1,
+                    'expected only one route to match "{}"; found {}'
+                    .format(string, [s[0].group(0) for s in matches]))
+            else:
+                self.assertTrue(
+                    len(matches) > 1,
+                    'expected more than one route to match "{}"; found {}'
+                    .format(string, [s[0].group(0) for s in matches]))
+
+            self.assertEqual(matches[0][0].group(0), string, 'failed for "{}"'.format(string))
+
+            if name:
+                expected = '_ex_route_' + name
+            else:
+                expected = '_ex_route_' + string
+
+            self.assertEqual(matches[0][2].__name__[0:len(expected)], expected)
+
+        # TODO [review] assert_route('g') works for the command scanner but fails a regex match, why?
+        # TODO [review] assert_route('global') works for the command scanner but fails a regex match, why?
+
+        assert_route('ab')
+        assert_route('abbreviate')
+        assert_route('bro')
+        assert_route('browse')
+        assert_route('buffers')
+        assert_route('cd')
+        assert_route('cdd')
+        assert_route('clo')
+        assert_route('close')
+        assert_route('co')
+        assert_route('copy')
+        assert_route('cq')
+        assert_route('cquit')
+        assert_route('d')
+        assert_route('delete')
+        assert_route('e')
+        assert_route('edit')
+        assert_route('exi', first=True)
+        assert_route('exit', first=True)
+        assert_route('f')
+        assert_route('file')
+        assert_route('files', first=True, name='buffers')
+        assert_route('h')
+        assert_route('help')
+        assert_route('let ', name='let')
+        assert_route('ls', name='buffers')
+        assert_route('m')
+        assert_route('move')
+        assert_route('new')
+        assert_route('nn')
+        assert_route('nnoremap')
+        assert_route('no')
+        assert_route('noremap')
+        assert_route('nun')
+        assert_route('nunmap')
+        assert_route('ono')
+        assert_route('onoremap')
+        assert_route('ou')
+        assert_route('ounmap')
+        assert_route('p')
+        assert_route('print')
+        assert_route('pw')
+        assert_route('pwd')
+        assert_route('q')
+        assert_route('qa')
+        assert_route('qall')
+        assert_route('quit')
+        assert_route('r')
+        assert_route('read')
+        assert_route('reg')
+        assert_route('registers')
+        assert_route('s')
+        assert_route('se')
+        assert_route('set')
+        assert_route('setl')
+        assert_route('setlocal')
+        assert_route('sh')
+        assert_route('shell')
+        assert_route('snor')
+        assert_route('snoremap')
+        assert_route('sp')
+        assert_route('split')
+        assert_route('substitute')
+        assert_route('sunm')
+        assert_route('sunmap')
+        assert_route('tabfir')
+        assert_route('tabfirst')
+        assert_route('tabl')
+        assert_route('tablast')
+        assert_route('tabn')
+        assert_route('tabnext')
+        assert_route('tabo')
+        assert_route('tabonly')
+        assert_route('tabp')
+        assert_route('tabprevious')
+        assert_route('tabr', name='tabfirst')
+        assert_route('tabrewind', name='tabfirst')
+        assert_route('una')
+        assert_route('unabbreviate')
+        assert_route('unm')
+        assert_route('unmap')
+        assert_route('vn')
+        assert_route('vnoremap')
+        assert_route('vs')
+        assert_route('vsplit')
+        assert_route('vu')
+        assert_route('vunmap')
+        assert_route('w')
+        assert_route('wa')
+        assert_route('wall')
+        assert_route('wq')
+        assert_route('wqall', first=True)
+        assert_route('write')
+        assert_route('x', name='exit')
+        assert_route('xa', name="wqall")
+        assert_route('xall', name="wqall")
+        assert_route('xit', name='exit')
+        assert_route('y')
+        assert_route('yank')